chore: separate model class from load yaml and toml packages
