+++ config/model_test.go
@@ -1,4 +1,4 @@
-package model
+package config
 
 import (
 	"testing"
